==> accidents/1 - data-capture.ipynb <==
## Prepare Data

import urllib.request
from tqdm import tqdm_notebook as tqdm
from tqdm import tqdm
from multiprocessing import Pool
import pandas as pd
import numpy as np
import json
import glob
import os

tqdm.pandas()

path = '/home/altieris/docker/jupyter/notebooks/ceabs/accidents/data/input/'
data_files = glob.glob(os.path.join(path, "*.csv"))  
num_partitions = 15 #number of partitions to split dataframe
num_cores = 7 #number of cores on your machine

df = pd.concat(map(lambda file: pd.read_csv(file, sep=';',encoding='latin-1'), data_files))

df.shape

#df.drop(['id','uop','delegacia','regional','sentido_via','uso_solo','ignorados'],inplace=True,axis=1)

df.head()

def concatenate_data(latitude,longitude):
    return "http://revgeo1.ceabsservicos.com:5000/search/json/"+str(latitude)+';'+str(longitude)

def work(x):
    x['revgeo_url'] = x.progress_apply(lambda row: concatenate_data(row['latitude'], row['longitude']), axis=1) 
    return x

def parallelize_dataframe(df, func):
    df_split = np.array_split(df, num_partitions)
    pool = Pool(num_cores)
    df = pd.concat(pool.map(func, df_split))
    pool.close()
    pool.join()
    return df

def call_url(url):
    try:
        contents = urllib.request.urlopen(url).read()
        wjdata = json.loads(contents)
        return wjdata['velocidadeMaximaVia_']
    except:
        return 0

    
def call_revgeo(data):
    data['velocidade_via'] = data.progress_apply(lambda row: call_url(row['revgeo_url']), axis=1) 
    return data

df  = parallelize_dataframe(df,work)

df = parallelize_dataframe(df, call_revgeo)

df.drop(['revgeo_url'],inplace=True,axis=1)

df.head()

df.shape

list(df)

path = '/home/altieris/docker/jupyter/notebooks/ceabs/accidents/data/input/raw/'
df.to_csv(path+'/accidents_brasil.csv', sep=';', encoding='utf-8',index=False)
