==> accidents/2 - data cleansing and Feature Engineering.ipynb <==
## Data Cleansing and Feature Selection

import random
import pandas as pd
import numpy as np
import datetime

%matplotlib inline

path = '/home/altieris/docker/jupyter/notebooks/ceabs/accidents/data/input/raw/'

df = pd.read_csv(path +'accidents_brasil.csv', sep=';')
df.head()

df.index = pd.to_datetime(df.data_inversa)
df['year'] = df.index.year
df['month'] = df.index.month
df['day'] = df.index.day

def regiao(uf):
    if uf in ('SP','RJ','MG','ES'):
        return 'SUDESTE'
    elif uf in ('AM','RR','AP','PA','TO','RO','AC'):
        return 'NORTE'
    elif uf in ('MA','PI','CE','RN','PE','PB','SE','AL','BA'):
        return 'NORDESTE'
    elif uf in ('MT','MS','GO','DF'):
        return 'CENTRO-OESTE'
    else:
        return 'SUL'

df['regiao'] = df.apply(lambda row: regiao(row['uf']), axis=1)

### Feriados 2017

| Data       	| Dia da semana    	|  Feriado                                  	|
|------------	|----------------	|--------------------------------------------	|
| 2017-1-1   	| domingo        	|  Confraternização Universal                	|
| 2017-2-27  	| segunda-feira  	| Carnaval                                   	|
| 2017-2-28  	| terça-feira    	| Carnaval                                   	|
| 2017-4-14  	| sexta-feira    	| Paixão de Cristo                           	|
| 2017-4-21  	| sexta-feira    	| Tiradentes                                 	|
| 2017-5-1   	| segunda-feira  	| Dia do Trabalho                            	|
| 2017-6-15  	| quinta-feira   	| Corpus Christi                             	|
| 2017-9-7   	| quinta-feira   	| Independência do Brasil                    	|
| 2017-10-12 	| quinta-feira   	| Nossa Sr.a Aparecida - Padroeira do Brasil 	|
| 2017-11-2  	| quinta-feira   	| Finados                                    	|
| 2017-11-15 	| quarta-feira   	| Proclamação da República                   	|
| 2017-12-25 	| segunda-feira  	| Natal                                      	|

### Feriados 2018

| Data       	| Dia da Semana 	| Feriado                                    	|
|------------	|---------------	|--------------------------------------------	|
| 2018-1-1   	| segunda-feira 	| Confraternização Universal                 	|
| 2018-2-12  	| segunda-feira 	| Carnaval                                   	|
| 2018-2-13  	| terça-feira   	| Carnaval                                   	|
| 2018-3-30  	| sexta-feira   	| Paixão de Cristo                           	|
| 2018-4-21  	| sábado        	| Tiradentes                                 	|
| 2018-5-1   	| terça-feira   	| Dia do Trabalho                            	|
| 2018-5-31  	| quinta-feira  	| Corpus Christi                             	|
| 2018-9-7   	| sexta-feira   	| Independência do Brasil                    	|
| 2018-10-12 	| sexta-feira   	| Nossa Sr.a Aparecida - Padroeira do Brasil 	|
| 2018-11-2  	| sexta-feira   	| Finados                                    	|
| 2018-11-15 	| quinta-feira  	| Proclamação da República                   	|
| 2018-12-25 	| terça-feira   	| Natal                                      	|

def feriado(data):
    if data in ('2017-01-01','2017-02-27','2017-02-28','2017-04-14','2017-04-21','2017-05-01','2017-06-15','2017-09-07','2017-10-12','2017-11-02','2017-11-15','2017-12-25'): # feriados 2017
        return 1
    elif data in ('2018-01-01','2018-02-12','2018-02-13','2018-03-30','2018-04-21','2018-05-01','2018-05-31','2018-09-07','2018-10-12','2018-11-02','2018-11-15','2018-12-25'): # feriados 2018
        return 1
    else:
        return 0


def desc_feriado(data):
    if data in ('2017-01-01','2018-01-01'):
        return 'Confraternização Universal'
    elif data in ('2017-02-27','2017-02-28','2018-02-12','2018-02-13'):
        return 'Carnaval'
    elif data in ('2017-04-14','2018-03-30'):
        return 'Paixão de Cristo'
    elif data in ('2017-04-21','2018-04-21'):
        return 'Tiradentes'
    elif data in ('2017-05-1','2018-05-1'):
        return 'Dia do Trabalho'
    elif data in ('2017-06-15','2018-05-31'):
        return 'Corpus Christi'
    elif data in ('2017-09-7','2018-09-7'):
        return 'Independência do Brasil'
    elif data in ('2017-10-12','2018-10-12'):
        return 'Nossa Sr.a Aparecida - Padroeira do Brasil'
    elif data in ('2017-11-02','2018-11-02'):
        return 'Finados'
    elif data in ('2017-11-15','2018-11-15'):
        return 'Proclamação da República'
    elif data in ('2017-12-25','2018-12-25'):
        return 'Natal'
    else:
        return ''
    
df['feriado'] = df.apply(lambda row: feriado(row['data_inversa']), axis=1)
df['nome_feriado'] = df.apply(lambda row: desc_feriado(row['data_inversa']), axis=1)

http://www.iag.usp.br/astronomia/inicio-das-estacoes-do-ano

### Estações do Ano 2017
Outono     2017-03-20  
Inverno    2017-06-21  
Primavera  2017-09-22  
Verão      2017-12-21  

### Estações do Ano 2018
Outono     2018-03-20  
Inverno    2018-06-21  
Primavera  2018-09-22  
Verão      2018-12-21  


def desc_estacao_ano(data):
    
    if (data >= '2016-12-21' and  data < '2017-03-20'):
        return 'verão'
    elif (data >= '2017-03-20' and data < '2017-06-21') or (data >= '2018-03-20' and data < '2018-06-21'):
        return 'outono'
    elif (data >= '2017-06-21' and data < '2017-09-22') or (data >= '2018-06-21' and data < '2018-09-22'):
        return 'inverno'
    elif (data >= '2017-09-22' and data < '2017-12-21')  or (data >= '2018-09-22' and data < '2018-12-21'):
        return 'primavera'
    elif (data >= '2017-12-21'  and data < '2018-03-20') or (data >= '2018-12-21' and data < '2019-03-20'):
        return 'verão'
    else:
        return ''          
                 
df['estacao'] = df.apply(lambda row: desc_estacao_ano(row['data_inversa']), axis=1)

df.tail()

# norte 5.271944, -60.2125
# sul -84, 79.494167
# leste -20.461944, -28.835833
# oeste -7.535833, -73.990556

df = df[(df['latitude'] > -90) & (df['latitude'] < 90)]
df = df[(df['longitude'] > -180) & (df['longitude'] < 180)]

sul = df[(df['regiao']  == 'SUL')]
sudeste = df[(df['regiao']  == 'SUDESTE')]
centro_oeste = df[(df['regiao']  == 'CENTRO-OESTE')]
norte = df[(df['regiao']  == 'NORTE')]
nordeste = df[(df['regiao']  == 'NORDESTE')]

sul = sul[(sul['latitude'] != 0)]
sul = sul[(sul['latitude'] != -2.70895468)]
sul = sul[(sul['latitude'] != -11.59843162)]
sul = sul[(sul['latitude'] != 26)]
sul = sul[(sul['latitude'] != -22.00960734)] 
sul = sul[(sul['latitude'] != -26)]  

sudeste = sudeste[(sudeste['latitude'] != -19)]
sudeste = sudeste[(sudeste['latitude'] != 19)]
sudeste = sudeste[(sudeste['latitude'] != 23)]
sudeste = sudeste[(sudeste['latitude'] != 0)]
sudeste = sudeste[(sudeste['latitude'] != -17.39257927)]

centro_oeste = centro_oeste[(centro_oeste['latitude'] != -9.10209674)] 
centro_oeste = centro_oeste[(centro_oeste['latitude'] != -5.0446853)] 
centro_oeste = centro_oeste[(centro_oeste['latitude'] != -5.05059064)]  
centro_oeste = centro_oeste[(centro_oeste['longitude'] != -42.7662325)]
centro_oeste = centro_oeste[(centro_oeste['longitude'] != -42.7662325)]

nordeste = nordeste[(nordeste['longitude'] != -94.49999571)] 
nordeste = nordeste[(nordeste['latitude'] < 0)] 
nordeste = nordeste[(nordeste['latitude'] != -19.31114336)] 
nordeste = nordeste[(nordeste['latitude'] != -85.0494645)] 

path = '/home/altieris/docker/jupyter/notebooks/ceabs/accidents/data/regioes/'

sul.to_csv(path+'accidents_sul.csv', sep=';', encoding='utf-8',index=False)

sudeste.to_csv(path+'accidents_sudeste.csv', sep=';', encoding='utf-8',index=False)


centro_oeste.to_csv(path+'accidents_centro_oeste.csv', sep=';', encoding='utf-8',index=False)


norte.to_csv(path+'accidents_norte.csv', sep=';', encoding='utf-8',index=False)


nordeste.to_csv(path+'accidents_nordeste.csv', sep=';', encoding='utf-8',index=False)

df1 = sul.append(sudeste)
df2 = df1.append(centro_oeste)
df3 = df2.append(norte)
df4 = df3.append(nordeste)
df4.shape

df4.to_csv(path+'accidents_brasil_cleansed.csv', sep=';', encoding='utf-8',index=False)
