==> accidents/3 - exploratory-data-analysis geral.ipynb <==
## Exploratory Data Analysis

import random
import seaborn as sns
import pandas as pd
import numpy as np
import datetime
import matplotlib.pyplot as plt

sns.set(style="whitegrid")

%matplotlib inline

path = '/home/altieris/docker/jupyter/notebooks/ceabs/accidents/data/'
brasil = pd.read_csv(path+'accidents_brasil.csv',sep=";")

br = brasil[(brasil['year']==2017)].groupby(['uf','month']).size().reset_index(name='counts')
#fig, ax = plt.subplots(figsize=(20,10))
sns.barplot(x='month',y='counts',data=br,hue="uf",orient='h')
#sns.catplot(x='month',y='counts',data=br,hue="uf",kind='count')
#a = sns.catplot(x="uf", hue="tipo_acidente", kind="count",col="year", data=br)

# qtd acidentes por dia da semana
# qtd acidentes por mês
# qtd acidentes por período do dia
# qtd acidentes por uf
# qtd acidentes por causa
# distribuição de mortos e feridps por tipo de acidente
# distribuição de mortos e feridos por estado
# qtd de acidentes por tipo e velocidade da via
# qtd de acidentes por BR
