==> accidents/3.2 - Sudeste.ipynb <==
### Sudeste


coords = sudeste[['latitude','longitude','causa_acidente','velocidade_via','data_inversa']]
fig = folium.Figure()
map_orgs = folium.Map(width=1500,height=1000,location=[-13, -51], zoom_start=5)
MarkerClusterScript(coords.to_json(orient="records"),callback=create_marker).add_to(map_orgs)
map_orgs.add_to(fig)

b = sns.catplot(x="month", hue="uf", kind="count",col="year", data=sudeste)
b.fig.set_size_inches(25,6)

sudeste_df = sudeste[(sudeste['uf'] == 'MG') & (sudeste['year'] == 2017)]
a = sns.catplot(x="condicao_metereologica", hue="causa_acidente", kind="count", data=sudeste_df,height=10,aspect=2)

sudeste_df = sudeste[(sudeste['uf'] == 'MG') & (sudeste['year'] == 2017)]

sd = sudeste_df[['tipo_acidente','causa_acidente']]
df.reset_index(drop = True, inplace = True)
a = sns.pairplot(sd)
