==> accidents/3.3 - Centro Oeste.ipynb <==
## Centro Oeste

coords = centro_oeste[['latitude','longitude','causa_acidente','velocidade_via','data_inversa']]
fig = folium.Figure()
map_orgs = folium.Map(width=1500,height=1000,location=[-13, -51], zoom_start=3)
MarkerClusterScript(coords.to_json(orient="records"),callback=create_marker).add_to(map_orgs)
map_orgs.add_to(fig)

centro_oeste.shape

centro_oeste.head()

list(centro_oeste)

#a4_dims = (11.7, 8.27)
#fig, ax = pyplot.subplots(nrows=6, ncols=5)
a = sns.catplot(x="month", hue="uf", kind="count",col="year", data=centro_oeste)
a.fig.set_size_inches(25,6)

a = sns.catplot(x="uf", hue="classificacao_acidente", kind="count",col="year", data=centro_oeste)
a.fig.set_size_inches(25,6)

a = sns.catplot(x="uf", hue="tipo_acidente", kind="count",col="year", data=centro_oeste)
a.fig.set_size_inches(35,6)

a = sns.catplot(x="tipo_pista", hue="uf", kind="count",col="year", data=centro_oeste)
a.fig.set_size_inches(25,6)

a = sns.catplot(x="tipo_pista", hue="tipo_acidente", kind="count",col="uf", data=centro_oeste)
a.fig.set_size_inches(30,6)

centro_oeste_df = sul[(centro_oeste['uf'] == 'PR') & (centro_oeste['year'] == 2017)]

a = sns.catplot(x="condicao_metereologica", hue="tipo_acidente", kind="count", data=centro_oeste_df,height=10,aspect=2)
#a.fig.set_size_inches(70,10)

centro_oeste_df = centro_oeste[(centro_oeste['uf'] == 'PR') & (centro_oeste['year'] == 2017)]
a = sns.catplot(x="condicao_metereologica", hue="causa_acidente", kind="count", data=centro_oeste_df,height=10,aspect=2)

centro_oeste_df = centro_oeste[(centro_oeste['uf'] == 'SC') & (centro_oeste['year'] == 2017)]
a = sns.catplot(x="condicao_metereologica", hue="causa_acidente", kind="count", data=centro_oeste_df,height=10,aspect=2)

centro_oeste_df = centro_oeste[(centro_oeste['uf'] == 'RS') & (centro_oeste['year'] == 2017)]
a = sns.catplot(x="condicao_metereologica", hue="causa_acidente", kind="count", data=centro_oeste_df,height=10,aspect=2)
