==> accidents/3.4 - Norte.ipynb <==
## Região Sul

coords = sul[['latitude','longitude','causa_acidente','velocidade_via','data_inversa']]
fig = folium.Figure()
map_orgs = folium.Map(width=1500,height=1000,location=[-13, -51], zoom_start=3)
MarkerClusterScript(coords.to_json(orient="records"),callback=create_marker).add_to(map_orgs)
map_orgs.add_to(fig)

sul.shape

sul.head()

list(sul)

#a4_dims = (11.7, 8.27)
#fig, ax = pyplot.subplots(nrows=6, ncols=5)
a = sns.catplot(x="month", hue="uf", kind="count",col="year", data=sul)
a.fig.set_size_inches(25,6)

a = sns.catplot(x="uf", hue="classificacao_acidente", kind="count",col="year", data=sul)
a.fig.set_size_inches(25,6)

a = sns.catplot(x="uf", hue="tipo_acidente", kind="count",col="year", data=sul)
a.fig.set_size_inches(35,6)

a = sns.catplot(x="tipo_pista", hue="uf", kind="count",col="year", data=sul)
a.fig.set_size_inches(25,6)

a = sns.catplot(x="tipo_pista", hue="tipo_acidente", kind="count",col="uf", data=sul)
a.fig.set_size_inches(30,6)

sul_df = sul[(sul['uf'] == 'PR') & (sul['year'] == 2017)]

a = sns.catplot(x="condicao_metereologica", hue="tipo_acidente", kind="count", data=sul_df,height=10,aspect=2)
#a.fig.set_size_inches(70,10)

sul_df = sul[(sul['uf'] == 'PR') & (sul['year'] == 2017)]
a = sns.catplot(x="condicao_metereologica", hue="causa_acidente", kind="count", data=sul_df,height=10,aspect=2)

sul_df = sul[(sul['uf'] == 'SC') & (sul['year'] == 2017)]
a = sns.catplot(x="condicao_metereologica", hue="causa_acidente", kind="count", data=sul_df,height=10,aspect=2)

sul_df = sul[(sul['uf'] == 'RS') & (sul['year'] == 2017)]
a = sns.catplot(x="condicao_metereologica", hue="causa_acidente", kind="count", data=sul_df,height=10,aspect=2)
