==> accidents/4 - ml-clustering.ipynb <==
### Clustering

 ---------------
|       |  /    |
| -1/1  | / 1/1 |
|       |/      |
|-------------- |
|      /|       |
|-1/-1/ | 1/-1  |
|    /  |       |
 ---------------

import json
import os
import numpy as np
import pandas as pd
from sklearn.cluster import DBSCAN
from shapely.geometry import Polygon,MultiPoint,Point,LineString
from makeclustermap import MarkerClusterScript
import shapely.geometry
import sys, getopt
import folium
import random
import seaborn as sns
import datetime
import matplotlib.pyplot as plt

%matplotlib inline

sns.set(style="whitegrid")

path = os.getcwd()
kms_per_radian = 6371.0088

# 100 metros
epsilon = 0.1 / kms_per_radian

def load_dataset(dataset_name,index_column_name,sep):
    try:
        dataframe = pd.read_csv(dataset_name, sep=sep, parse_dates=True, infer_datetime_format=True)
        #dataframe.index = pd.to_datetime(dataframe[index_column_name])
        return dataframe
    except Exception:
        print("Ocorreu um erro ao tentar carregar o arquivo.") 

def dbscan_fit(coords_dataset, min_samples=3):
    model = DBSCAN(eps=epsilon, min_samples=min_samples, algorithm='ball_tree', metric='haversine', n_jobs=1).fit(np.radians(coords_dataset))
    num_clusters = len(set(model.labels_))
    print('Number of clusters: {:,}'.format(num_clusters))
    return model

df = load_dataset(path+'/data/regioes/accidents_brasil_cleansed.csv','data_inversa',';')    
df[['latitude','longitude']].head()

coords = df.drop_duplicates(['longitude','latitude'])[['latitude','longitude']]#.sort_values(by=['latitude','longitude'])

coords = coords.sort_values(by=['latitude', 'longitude'], ascending=False)

coords.as_matrix(columns=['longitude', 'latitude'])

model = dbscan_fit(coords)

model.labels_[0:100] # Firt 100 points

coords[model.labels_ == 100] # pontos pentencentes ao 200º cluster

clusters = pd.Series([coords[model.labels_ == n] for n in range(len(set(model.labels_)))]) # cria grupos com os pontos 

len(clusters)

c = [[i,len(clusters[i])] for i in range(0,len(clusters))]
clusters_len = pd.DataFrame(c, columns=('index','tamanho')) 

clusters_len.sort_values(by=['tamanho'],ascending=False).head(10)

clusters_len = clusters_len[(clusters_len.tamanho > 10 )].sort_values(by=['tamanho'],ascending=False)

clusters_len.head()

clusters_len[clusters_len['tamanho'] == clusters_len['tamanho'].max()]# cluster com maior numero de pontos
idx = clusters_len[clusters_len['tamanho'] == clusters_len['tamanho'].max()].index[0]

c = pd.DataFrame(clusters[73], columns=('longitude','latitude'))
#c = c.append(pd.DataFrame(clusters[107], columns=('longitude','latitude')))

def create_marker_clustered(row, popup=None):
    """Returns a L.marker object"""
    icon = L.AwesomeMarkers.icon({markerColor: row.color})    
    marker = L.marker(L.LatLng(row.latitude, row.longitude))
    marker.setIcon(icon)
    marker.bindPopup(row.latitude + ' ' + row.longitude + ' ocorrencias:'+row.nr_ocurrences)
    return marker


width, height = '100%', 600

#coords = clustered_data[['longitude','latitude','nr_ocurrences']]

coords = c                     

fig = folium.Figure(width=width, height=height)
map_orgs = folium.Map(width=width,height=height,location=[-12.032378, -51.574553], zoom_start=4)
MarkerClusterScript(coords.to_json(orient="records"),callback=create_marker_clustered).add_to(map_orgs)
map_orgs.add_to(fig)

cluster0 = pd.DataFrame(clusters[73],columns=['longitude','latitude'])
cluster0 = cluster0.set_index(['longitude','latitude'])

#df = df.set_index(['longitude','latitude'])

cluster0.head()

clusters[73][:10]#[0][1]

import urllib.request
from tqdm import tqdm_notebook as tqdm
from tqdm import tqdm
from multiprocessing import Pool
import pandas as pd
import numpy as np
import json
import glob
import os

tqdm.pandas()

def concatenate_data(df):
    
    try:
        call_url_ws = "http://10.3.0.105:5000/route/v1/driving/"
        call_url_maps = "http://10.3.0.105:9966/?"
        for row in df.itertuples():
            call_url_ws = call_url_ws + str(row.longitude)+","+str(row.latitude)+';'
            call_url_maps = call_url_maps + '&loc='+str(row.latitude)+","+str(row.longitude)
        
        print(call_url_ws)
        print(call_url_maps)

        contents = urllib.request.urlopen(call_url_ws[:-1]).read()
        wjdata = json.loads(contents)    
        call_url_maps = "http://10.3.0.105:9966/?"
        for el in wjdata['waypoints']:
            call_url_maps = call_url_maps + '&loc='+str(el['location'][1])+","+str(el['location'][0])
        print(call_url_maps)
        print('-------------------------------------------------')
    except:
        pass
    
#clusters.apply(lambda row: concatenate_data(row)) 

import urllib.request

url = "http://127.0.0.1:5000/route/v1/driving/-48.35989058,-1.36535118;-48.35697234,-1.365378;-48.35748196,-1.365378;-48.35716009,-1.36540481"

contents = urllib.request.urlopen(url).read()
wjdata = json.loads(contents)    

call_url_maps = "http://localhost:9966/?"
for el in wjdata['waypoints']:
    call_url_maps = call_url_maps + '&loc='+str(el['location'][1])+","+str(el['location'][0])
    
print(call_url_maps)

clusters[73]

t = pd.DataFrame([])
matrix = clusters[100].as_matrix(columns=['longitude', 'latitude'])
for i in range(0,len(matrix)):
    t = t.append(df.loc[(df['longitude'] == matrix[i][0]) & (df['latitude'] == matrix[i][1])])

t.drop(['id','nome_feriado','estacao','year','month','day','data_inversa','uf','br','km','feriado','velocidade_via','regiao','uop','delegacia','regional','sentido_via','horario','dia_semana','tracado_via','uso_solo'],axis=1)

#coords['radius'] = (coords['nr_ocurrences'] - coords.nr_ocurrences.min()) / (coords.nr_ocurrences.max() - coords.nr_ocurrences.min()) * 200

class Cluster(object):

    def __init__(self,latitude_centroid,longitude_centroid,cluster_dataframe):
        self.latitude_centroid = latitude_centroid
        self.longitude_centroid = longitude_centroid
        self.cluster_dataframe = cluster_dataframe
        self.nr_points = len(cluster_dataframe)
        self.polygon = Polygon(cluster_dataframe.as_matrix(columns=['longitude', 'latitude']))

## ----------------------------------------------------------------------------------

# def get_centermost_point(cluster):
#     coords = cluster.sort_values(by=['latitude', 'longitude'], ascending=False)
#     coords = cluster.as_matrix(columns=['longitude', 'latitude'])
#     centroid = (MultiPoint(coords).centroid.x, MultiPoint(coords).centroid.y)
#     longitude_centroid = MultiPoint(coords).centroid.x
#     latitude_centroid = MultiPoint(coords).centroid.y
#     item_cluster = Cluster(latitude_centroid,longitude_centroid, cluster)
#     return item_cluster


# def generate_data(cluster_areas):
#     d = []
#     for i in range(len(cluster_areas)):
#         try:
#             polygon = Polygon(cluster_areas[i].points)
#             centroid = Point(cluster_areas[i].centroid)
#             x = polygon.buffer(buffer)
#             s = x.simplify(simplify_const, preserve_topology=True)
#             p = LineString(s.exterior.coords)
#             d.append((cluster_areas[i].latitude,cluster_areas[i].longitude,centroid, cluster_areas[i].nr_points(), polygon,  json.dumps(shapely.geometry.mapping(p))))
            
#         except Exception as inst:
#             pass
#     return d

x = get_centermost_point(coords)

x.cluster_dataframe#.polygon.coords

buffer = 0.008
simplify_const = 0.5


# y = x.polygon.buffer(buffer)
# s = y.simplify(simplify_const, preserve_topology=True)
# p = LineString(s.exterior.coords)
# print(p.wkt)
# json.dumps(shapely.geometry.mapping(p))

x.polygon.bounds

x.polygon.wkt

s.exterior.coords.xy

coords.head()

width, height = '100%', 600

colors = {'A' : 'red', 'B' : 'blue'}

map_osm = folium.Map(width=width,height=height,location=[-12.032378, -51.574553], zoom_start=4)

coords.head(1).apply(lambda row: folium.CircleMarker(location=[row['latitude'], row['longitude']],fill=True,fill_opacity=0.3,color = 'red', popup=str(row.nr_ocurrences),radius=row['radius'])
                                             .add_to(map_osm), axis=1)

map_osm

dataframe.to_csv(filename, index=False)

coords_ocurrences = sudeste.as_matrix(columns=['longitude', 'latitude'])
outputfile = 'accidents_clustered.csv'
generate_cluster(outputfile, coords_ocurrences)

#https://github.com/qingkaikong/blog/tree/master/28_DBSCAN
#https://www.aaai.org/Papers/KDD/1996/KDD96-037.pdf
#ftp://public.dhe.ibm.com/software/analytics/spss/documentation/modeler/17.1/br_po/ModelerCRISPDM.pdf
#https://exame.abril.com.br/brasil/as-10-rodovias-federais-mais-perigosas-do-brasil/
#http://www.dbs.ifi.lmu.de/Publikationen/Papers/KDD-96.final.frame.pdf
#https://arthur-e.github.io/Wicket/sandbox-gmaps3.html
# https://www.jamesrcroft.com/2015/06/snapping-gps-tracks-to-roads/
